05-keyword-extractor/requirements.txt
openai
python-dotenv
streamlit
pydantic
